+++ src/lib.rs
@@ -138,5 +138,6 @@ impl DustMasker {
 #[pymodule]
 fn _pydustmasker(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<DustMasker>()?;
+    m.add("__version__", env!("CARGO_PKG_VERSION"))?;
     Ok(())
 }